pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "llmvcr"
version = "0.1.0"
description = "Record and replay LLM API responses for fast, free, deterministic tests"
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.8"
keywords = ["llm", "testing", "openai", "anthropic", "vcr", "cassette", "mock"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Testing",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "pyyaml>=6.0",
    "wrapt>=1.14",
]

[project.optional-dependencies]
openai = ["openai>=1.0"]
anthropic = ["anthropic>=0.20"]
gemini = ["google-genai>=1.0"]
ollama = ["ollama>=0.2"]
groq = ["groq>=0.9"]
all = ["openai>=1.0", "anthropic>=0.20", "google-genai>=1.0", "ollama>=0.2", "groq>=0.9"]
dev = [
    "pytest>=7.0",
    "openai>=1.0",
    "anthropic>=0.20",
    "python-dotenv>=1.0",
]

[project.urls]
Homepage = "https://github.com/shahid-43/llmvcr"
Documentation = "https://github.com/shahid-43/llmvcr#readme"
Issues = "https://github.com/shahid-43/llmvcr/issues"

[project.scripts]
llmvcr = "llmvcr.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/llmvcr"]